인증 및 권한 체크하기

+++ src/main/java/shop/mtcoding/blog/board/BoardController.java
@@ -9,6 +9,8 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import shop.mtcoding.blog._core.errors.exception.Exception403;
+import shop.mtcoding.blog._core.errors.exception.Exception404;
 import shop.mtcoding.blog.user.User;
 
 import java.lang.annotation.Native;
@@ -24,6 +26,11 @@ public class BoardController {
     // @Transactional 트랜잭션 시간이 너무 길어져서 service에 넣어야함
     @PostMapping("/board/{id}/update")
     public String update(@PathVariable Integer id, BoardRequest.UpdateDTO reqDTO) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        Board board = boardRepository.findById(id);
+        if (sessionUser.getId() != board.getUser().getId()) {
+            throw new Exception403("게시글을 수정할 권한이 없습니다");
+        }
         boardRepository.updateById(id, reqDTO.getTitle(), reqDTO.getContent());
         return "redirect:/board/" + id;
     }
@@ -31,12 +38,21 @@ public class BoardController {
     @GetMapping("/board/{id}/update-form")
     public String updateForm(@PathVariable(name = "id") Integer id, HttpServletRequest request) {
         Board board = boardRepository.findById(id);
+
+        if (board == null) {
+            throw new Exception404("해당 게시글을 찾을 수 없습니다");
+        }
         request.setAttribute("board", board);
         return "/board/update-form"; // 서버가 내부적으로 index를 요청 - 외부에서는 다이렉트 접근이 안됨
     }
 
     @PostMapping("/board/{id}/delete")
     public String delete(@PathVariable Integer id) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        Board board = boardRepository.findById(id);
+        if (sessionUser.getId() != board.getUser().getId()) {
+            throw new Exception403("게시글을 삭제할 권한이 없습니다");
+        }
         boardRepository.deleteById(id);
         return "redirect:/";
     }
@@ -61,8 +77,18 @@ public class BoardController {
     }
 
     @GetMapping("/board/{id}")
-    public String detail(@PathVariable Integer id, HttpServletRequest request) { // Integer : 없으면 null, int : 0
+    public String detail(@PathVariable Integer id, HttpServletRequest request) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
         Board board = boardRepository.findByIdJoinUser(id);
+
+        // 로그인을 하고, 게시글의 주인이면 isOwner가 true가 된다.
+        boolean isOwner = false;
+        if(sessionUser != null){
+            if(sessionUser.getId() == board.getUser().getId()){
+                isOwner = true;
+            }
+        }
+        request.setAttribute("isOwner", isOwner);
         request.setAttribute("board", board);
         return "board/detail";
     }
